sqlite_to_postgres/requirements.txt
typing-extensions==3.10.0.2
python-dotenv==0.21.0
psycopg2==2.9.5
pytest==6.2.5
python-dateutil==2.8.2
